pyproject.toml
[tool.poetry]
name = "yoyo-cloud"
version = "0.1.0"
description = "Run SQL migrations stored in the cloud."
authors = ["Fasih Khatib <hellofasih.confound928@passinbox.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
yoyo-migrations = "^8.2.0"
setuptools = "^69.2.0"
s3fs = "^2024.3.1"


[tool.poetry.group.dev.dependencies]
ipython = "^8.22.2"


[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
tox = "^4.14.2"
testcontainers = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
